fix more types

+++ src/include/api/ivf_pq_index.h
@@ -580,7 +580,7 @@ class IndexIVFPQ {
   };
 
   // clang-format off
-  using constructor_function = std::function<std::unique_ptr<index_base>(size_t, uint32_t, uint64_t, float, float, std::optional<TemporalPolicy>, DistanceMetric, uint64_t)>;
+  using constructor_function = std::function<std::unique_ptr<index_base>(size_t, uint32_t, uint32_t, float, float, std::optional<TemporalPolicy>, DistanceMetric, uint64_t)>;
   using table_type = std::map<std::tuple<tiledb_datatype_t, tiledb_datatype_t, tiledb_datatype_t>, constructor_function>;
   static const table_type dispatch_table;
